Do not wipe index when calling remove on new record - closes #405
(cherry picked from commit 381bab0748f8e9be328672eb4589ac0376bc909d)

+++ lib/searchkick/index.rb
@@ -59,11 +59,14 @@ module Searchkick
     end
 
     def remove(record)
-      client.delete(
-        index: name,
-        type: document_type(record),
-        id: search_id(record)
-      )
+      id = search_id(record)
+      unless id.blank?
+        client.delete(
+          index: name,
+          type: document_type(record),
+          id: id
+        )
+      end
     end
 
     def import(records)
